+++ src/ShellCheck/Analytics.hs
@@ -3159,11 +3159,11 @@ checkNullaryExpansionTest params t =
         TC_Nullary _ _ word ->
             case getWordParts word of
                 [t] | isCommandSubstitution t ->
-                    styleWithFix id 2243 "Prefer explicit -n to check for output (or run command without [/[[ to check for success)." fix
+                    verboseWithFix id 2243 "Prefer explicit -n to check for output (or run command without [/[[ to check for success)." fix
 
                 -- If they're constant, you get SC2157 &co
                 x | all (not . isConstant) x ->
-                    styleWithFix id 2244 "Prefer explicit -n to check non-empty string (or use =/-ne to check boolean/integer)." fix
+                    verboseWithFix id 2244 "Prefer explicit -n to check non-empty string (or use =/-ne to check boolean/integer)." fix
                 _ -> return ()
             where
                 id = getId word