+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - SC2247: Warn about $"(cmd)" and $"{var}"
 - SC2246: Warn if a shebang's interpreter ends with /
 - SC2245: Warn that Ksh ignores all but the first glob result in `[`
-- SC2243/SC2244: Suggest using explicit -n for `[ $foo ]`
+- SC2243/SC2244: Suggest using explicit -n for `[ $foo ]` (verbose)
 
 ### Changed
 - If a directive or shebang is not specified, a `.bash/.bats/.dash/.ksh`